print --regex tip when target is built during a target build error debounce timeout
this indicates the watcher process probably rebuilt ( wrote to disk ) the target bundle during an error.

This can happen sometimes when the watcher process embeds the error information directly inside the bundle.

miru by default treats a target bundle change as a successful build, so in these cases the -r regex watch flag can be used to identify when a target is successfully built.

This is more common these days with things like browserify and webpack.
They usually embed a `console.error(...)` or `throw new Error(...)` with
some information on the error that happened during the bundle process
instead of only printing to the terminal as they used to.